Compiles and runs all tests successfully on Windows
Changed from using the 'unix' library to 'unit-compat'. This caused
some compile error related to FileOffset that I suppressed by hiding
the FileOffset that was exported from Snap.Iteratee. I'm not sure if
this was the right thing to do or not but snap-server now compiles
and the test suite runs successfully.