Desperation tactic: try to force the garbage collector not to kill the SSLContext before we're finished with it.
The hypothesis here is that the SSLContext (which has an MVar finalizer
attached to it) is getting GCed before all of the uses are finished with it,
which is causing our testsuite segfault-on-exit. If this hypothesis is correct,
then touching the context on session shutdown should force the context to
remain live.