Initial unix socket support implementation
To use unix sockets specify "unix:/abs/path.sock" as address in config.
It is probably better to use separate parameter unix socket
specification. Server then can listen on http, https and unix sockets at
the same time.

Setting access permissions on socket is important if your snap server
and proxy, like nginx, are run with different users. So you can set
group write permission and add nginx user to same group as snap server.

On command line access rights are specified in octal.

Permissions are set with umask, so there is a race condition because
umask is process wide shared attribute. Setting permissions using
setFilePermissions can lead to race conditions if socket file is renamed
or unlinked while program is running.
In theory on Linux at least you can set permissions for socket Fd before
binding, but in my case it was setting not all permissions. Possibly
because it is also affected by umask set somewhere else.

Checking for HAS_SENDFILE, because it is defined on *nixes.

Using abs path only is a security feature, and is often recommended, for
example in TLPI.