Reject HTTP-versions not conforming to RFC 7230
RFC 7230 section 3.1.1 defines the first line of a request as

    request-line = method SP request-target SP HTTP-version CRLF

and RFC 7230 section 2.6 defines the HTTP-version part as

    HTTP-version = HTTP-name "/" DIGIT "." DIGIT
    HTTP-name    = %x48.54.54.50 ; "HTTP", case-sensitive

in other words, HTTP-version is intended to be two single digits.

The two most common values encountered are HTTP/1.0 and HTTP/1.1 and
so we provide direct matching for those.

This implicitly also fixes the bug of absurdly big version numbers
wrapping over into negative HTTP version integer parts.